Move reading snapshots to driver
