+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/snapshot/SnapshotsDriver.java
@@ -15,9 +15,10 @@
  */
 package sleeper.systemtest.dsl.snapshot;
 
+import sleeper.core.partition.PartitionTree;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
-import sleeper.core.statestore.transactionlog.TransactionLogSnapshot;
+import sleeper.core.statestore.AllReferencesToAllFiles;
 
 import java.util.Optional;
 
@@ -28,13 +29,13 @@ public interface SnapshotsDriver {
 
     void disableCreation(InstanceProperties instanceProperties);
 
-    default Optional<TransactionLogSnapshot> loadLatestFilesSnapshot(InstanceProperties instanceProperties, TableProperties tableProperties) {
+    default Optional<AllReferencesToAllFiles> loadLatestFilesSnapshot(InstanceProperties instanceProperties, TableProperties tableProperties) {
         throw new UnsupportedOperationException(
                 "Requested loading files snapshot for instance " + instanceProperties.get(ID) +
                         ", table " + tableProperties.getStatus() + ", not currently implemented for this driver");
     }
 
-    default Optional<TransactionLogSnapshot> loadLatestPartitionsSnapshot(InstanceProperties instanceProperties, TableProperties tableProperties) {
+    default Optional<PartitionTree> loadLatestPartitionsSnapshot(InstanceProperties instanceProperties, TableProperties tableProperties) {
         throw new UnsupportedOperationException(
                 "Requested loading partitions snapshot for instance " + instanceProperties.get(ID) +
                         ", table " + tableProperties.getStatus() + ", not currently implemented for this driver");
